+++ contracts/staking/NftStaking.sol
@@ -18,14 +18,17 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
     using SafeCast for uint256;
     using SafeMath for uint256;
     using SignedSafeMath for int256;
+    using ERC1155721TransferFallback for IERC1155721Transferrable;
 
     event RewardsAdded(uint256 startPeriod, uint256 endPeriod, uint256 rewardsPerCycle);
 
     event Started();
 
     event NftStaked(address staker, uint256 cycle, uint256 tokenId, uint256 weight);
+    event NftsBatchStaked(address staker, uint256 cycle, uint256[] tokenIds, uint256[] weights);
 
     event NftUnstaked(address staker, uint256 cycle, uint256 tokenId, uint256 weight);
+    event NftsBatchUnstaked(address staker, uint256 cycle, uint256[] tokenIds, uint256[] weights);
 
     event RewardsClaimed(address staker, uint256 cycle, uint256 startPeriod, uint256 periods, uint256 amount);
 
@@ -80,7 +83,7 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
     uint256 public startTimestamp;
 
     IERC20 public immutable rewardsTokenContract;
-    IWhitelistedNftContract public immutable whitelistedNftContract;
+    IERC1155721Transferrable public immutable whitelistedNftContract;
 
     uint32 public immutable cycleLengthInSeconds;
     uint16 public immutable periodLengthInCycles;
@@ -132,7 +135,7 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
     constructor(
         uint32 cycleLengthInSeconds_,
         uint16 periodLengthInCycles_,
-        IWhitelistedNftContract whitelistedNftContract_,
+        IERC1155721Transferrable whitelistedNftContract_,
         IERC20 rewardsTokenContract_
     ) internal {
         require(cycleLengthInSeconds_ >= 1 minutes, "NftStaking: invalid cycle length");
@@ -232,6 +235,10 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
 
     /*                                             ERC1155TokenReceiver                                             */
 
+    /**
+     * ERC1155Receiver hook for single transfer.
+     * @dev Reverts if the caller is not the whitelisted NFT contract.
+     */
     function onERC1155Received(
         address, /*operator*/
         address from,
@@ -239,10 +246,15 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
         uint256, /*value*/
         bytes calldata /*data*/
     ) external virtual override returns (bytes4) {
+        require(address(whitelistedNftContract) == msg.sender, "NftStaking: contract not whitelisted");
         _stakeNft(id, from);
         return _ERC1155_RECEIVED;
     }
 
+    /**
+     * ERC1155Receiver hook for batch transfer.
+     * @dev Reverts if the caller is not the whitelisted NFT contract.
+     */
     function onERC1155BatchReceived(
         address, /*operator*/
         address from,
@@ -250,9 +262,8 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
         uint256[] calldata, /*values*/
         bytes calldata /*data*/
     ) external virtual override returns (bytes4) {
-        for (uint256 i = 0; i < ids.length; ++i) {
-            _stakeNft(ids[i], from);
-        }
+        require(address(whitelistedNftContract) == msg.sender, "NftStaking: contract not whitelisted");
+        _batchStakeNfts(ids, from);
         return _ERC1155_BATCH_RECEIVED;
     }
 
@@ -273,7 +284,7 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
     function unstakeNft(uint256 tokenId) external virtual {
         TokenInfo memory tokenInfo = tokenInfos[tokenId];
 
-        require(tokenInfo.owner == msg.sender, "NftStaking: token not staked or incorrect token owner");
+        require(tokenInfo.owner == msg.sender, "NftStaking: not staked for owner");
 
         uint16 currentCycle = _getCycle(now);
 
@@ -294,14 +305,66 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
             tokenInfos[tokenId] = tokenInfo;
         }
 
-        try whitelistedNftContract.safeTransferFrom(address(this), msg.sender, tokenId, 1, "")  {} catch {
-            // the above call to the ERC1155 safeTransferFrom() function may fail if the recipient
-            // is a contract wallet not implementing the ERC1155TokenReceiver interface
-            // if this happens, the transfer falls back to a call to the ERC721 (non-safe) transferFrom function.
-            whitelistedNftContract.transferFrom(address(this), msg.sender, tokenId);
-        }
+        whitelistedNftContract.safeTransferFromWithFallback(address(this), msg.sender, tokenId, 1, "");
 
         emit NftUnstaked(msg.sender, currentCycle, tokenId, tokenInfo.weight);
+    }
+
+    /**
+     * Unstakes a batch of deposited NFTs from the contract.
+     * @dev Reverts if `tokenIds` is empty.
+     * @dev Reverts if the caller is not the original owner of any of the NFTs.
+     * @dev While the contract is enabled, reverts if any NFT is being unstaked before its staking freeze duration has elapsed.
+     * @dev While the contract is enabled, creates any missing snapshots, up-to the current cycle.
+     * @dev While the contract is enabled, emits the HistoriesUpdated event.
+     * @dev Emits the NftsBatchUnstaked event for each NFT unstaked.
+     * @param tokenIds The token identifiers, referencing the NFTs being unstaked.
+     */
+    function batchUnstakeNfts(uint256[] calldata tokenIds) external {
+        uint256 numTokens = tokenIds.length;
+        require(numTokens != 0, "NftStaking: no tokens");
+
+        uint16 currentCycle = _getCycle(now);
+        int128 totalUnstakedWeight = 0;
+        uint256[] memory values = new uint256[](numTokens);
+        uint256[] memory weights = new uint256[](numTokens);
+
+        for (uint256 index = 0; index < numTokens; ++index) {
+            uint256 tokenId = tokenIds[index];
+
+            TokenInfo memory tokenInfo = tokenInfos[tokenId];
+
+            require(tokenInfo.owner == msg.sender, "NftStaking: not staked for owner");
+
+            if (enabled) {
+                // ensure that at least an entire cycle has elapsed before
+                // unstaking the token to avoid an exploit where a a fukll cycle
+                // would be claimable if staking just before the end of a cycle
+                // and unstaking right after the start of the new cycle
+                require(currentCycle - tokenInfo.depositCycle >= 2, "NftStaking: token still frozen");
+
+                // clear the token owner to ensure it cannot be unstaked again
+                // without being re-staked
+                tokenInfos[tokenId].owner = address(0);
+
+                // we can use unsafe math here since the maximum total staked
+                // weight that a staker can unstake must fit within uint128
+                // (i.e. the staker snapshot stake limit)
+                uint64 weight = tokenInfo.weight;
+                totalUnstakedWeight -= weight; // this is safe
+                weights[index] = weight;
+            }
+
+            values[index] = 1;
+        }
+
+        emit NftsBatchUnstaked(msg.sender, currentCycle, tokenIds, weights);
+
+        if (enabled) {
+            _updateHistories(msg.sender, totalUnstakedWeight, currentCycle);
+        }
+
+        whitelistedNftContract.safeBatchTransferFromWithFallback(address(this), msg.sender, tokenIds, values, "");
     }
 
     /**
@@ -422,7 +485,6 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
 
     /**
      * Stakes the NFT received by the contract for its owner. The NFT's weight will count for the current cycle.
-     * @dev Reverts if the caller is not the whitelisted NFT contract.
      * @dev Emits an HistoriesUpdated event.
      * @dev Emits an NftStaked event.
      * @param tokenId Identifier of the staked NFT.
@@ -453,6 +515,44 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
         tokenInfos[tokenId] = TokenInfo(tokenOwner, weight, currentCycle, 0);
 
         emit NftStaked(tokenOwner, currentCycle, tokenId, weight);
+    }
+
+    /**
+     * Stakes the NFT received by the contract for its owner. The NFT's weight will count for the current cycle.
+     * @dev Reverts if `tokenIds` is empty.
+     * @dev Emits an HistoriesUpdated event.
+     * @dev Emits an NftStaked event.
+     * @param tokenIds Identifiers of the staked NFTs.
+     * @param tokenOwner Owner of the staked NFTs.
+     */
+    function _batchStakeNfts(uint256[] memory tokenIds, address tokenOwner) internal isEnabled hasStarted {
+        uint256 numTokens = tokenIds.length;
+        require(numTokens != 0, "NftStaking: no tokens");
+
+        uint16 currentCycle = _getCycle(now);
+        uint128 totalStakedWeight = 0;
+        uint256[] memory weights = new uint256[](numTokens);
+
+        for (uint256 index = 0; index < numTokens; ++index) {
+            uint256 tokenId = tokenIds[index];
+            require(currentCycle != tokenInfos[tokenId].withdrawCycle, "NftStaking: unstaked token cooldown");
+            uint64 weight = _validateAndGetNftWeight(tokenId);
+            totalStakedWeight += weight; // This is safe
+            weights[index] = weight;
+            tokenInfos[tokenId] = TokenInfo(tokenOwner, weight, currentCycle, 0);
+        }
+
+        _updateHistories(tokenOwner, int128(totalStakedWeight), currentCycle);
+
+        // initialise the next claim if it was the first stake for this staker or if
+        // the next claim was re-initialised (ie. rewards were claimed until the last
+        // staker snapshot and the last staker snapshot has no stake)
+        if (nextClaims[tokenOwner].period == 0) {
+            uint16 currentPeriod = _getPeriod(currentCycle, periodLengthInCycles);
+            nextClaims[tokenOwner] = NextClaim(currentPeriod, uint64(globalHistory.length - 1), 0);
+        }
+
+        emit NftsBatchStaked(tokenOwner, currentCycle, tokenIds, weights);
     }
 
     /**
@@ -714,9 +814,33 @@ abstract contract NftStaking is ERC1155TokenReceiver, Ownable {
 }
 
 /**
- * @notice Interface for the NftStaking whitelisted NFT contract.
+ * @notice Interface for transferring 1155 and/or 721 NFTs.
  */
-interface IWhitelistedNftContract {
+interface IERC1155721Transferrable {
+    /**
+     * @notice Transfers `values` amount(s) of `ids` from the `from` address to the `to` address specified (with safety call).
+     * @dev Caller must be approved to manage the tokens being transferred out of the `from` account (see "Approval" section of the standard).
+     * MUST revert if `to` is the zero address.
+     * MUST revert if length of `ids` is not the same as length of `values`.
+     * MUST revert if any of the balance(s) of the holder(s) for token(s) in `ids` is lower than the respective amount(s) in `values` sent to the recipient.
+     * MUST revert on any other error.
+     * MUST emit `TransferSingle` or `TransferBatch` event(s) such that all the balance changes are reflected (see "Safe Transfer Rules" section of the standard).
+     * Balance changes and events MUST follow the ordering of the arrays (_ids[0]/_values[0] before _ids[1]/_values[1], etc).
+     * After the above conditions for the transfer(s) in the batch are met, this function MUST check if `to` is a smart contract (e.g. code size > 0). If so, it MUST call the relevant `ERC1155TokenReceiver` hook(s) on `to` and act appropriately (see "Safe Transfer Rules" section of the standard).
+     * @param from Source address
+     * @param to Target address
+     * @param ids IDs of each token type (order and length must match _values array)
+     * @param values Transfer amounts per token type (order and length must match _ids array)
+     * @param data Additional data with no specified format, MUST be sent unaltered in call to the `ERC1155TokenReceiver` hook(s) on `to`
+     */
+    function safeBatchTransferFrom(
+        address from,
+        address to,
+        uint256[] calldata ids,
+        uint256[] calldata values,
+        bytes calldata data
+    ) external;
+
     /**
      * ERC1155: Transfers `value` amount of an `id` from  `from` to `to` (with safety call). 
      * @dev Caller must be approved to manage the tokens being transferred out of the `from` account (see "Approval" section of the standard).
@@ -752,4 +876,44 @@ interface IWhitelistedNftContract {
         address to,
         uint256 tokenId
     ) external;
+}
+
+
+/**
+ * @title ERC1155721TransferFallback
+ * Library used to fall back on ERC721 non-safe transfer(s)
+ * in case of ERC1155 safe transfer failure. A failure can be
+ * caused by a contract-based wallet not implementing the
+ * ERC1155Receiver interface.
+ */
+library ERC1155721TransferFallback {
+    function safeBatchTransferFromWithFallback(
+        IERC1155721Transferrable self,
+        address from,
+        address to,
+        uint256[] memory ids,
+        uint256[] memory values,
+        bytes memory data
+    ) internal {
+        try self.safeBatchTransferFrom(from, to, ids, values, data) {
+        } catch {
+            uint256 length = ids.length;
+            for (uint256 i = 0; i < length; ++i) {
+                self.transferFrom(from, to, ids[i]);
+            }
+        }
+    }
+
+    function safeTransferFromWithFallback(
+        IERC1155721Transferrable self,
+        address from,
+        address to,
+        uint256 id,
+        uint256 value,
+        bytes memory data
+    ) internal {
+        try self.safeTransferFrom(from, to, id, value, data)  {} catch {
+            self.transferFrom(from, to, id);
+        }
+    }
 }