+++ test/contracts/staking/behaviors/staking.behavior.js
@@ -1,7 +1,13 @@
+const {accounts} = require('@openzeppelin/test-environment');
 const {BN, expectRevert, expectEvent} = require('@openzeppelin/test-helpers');
 const {constants} = require('@animoca/ethereum-contracts-core_library');
 
 const {debugCurrentState} = require('./debug.behavior');
+
+const {RarityWeights} = require('../constants');
+const TokenHelper = require('../../../utils/tokenHelper');
+
+const [creator] = accounts;
 
 const shouldSetNextClaimIfUnset = async function (staker, stateBefore, stateAfter) {
     if (stateBefore.lastGlobalSnapshotIndex.eq(new BN('-1'))) {
@@ -12,7 +18,7 @@ const shouldSetNextClaimIfUnset = async function (staker, stateBefore, stateAfte
     }
 };
 
-const shouldUpdateHistory = async function (receipt, eventName, staker, tokenIds, stateBefore, stateAfter) {
+const shouldUpdateHistory = async function (receipt, action, staker, tokenIds, stateBefore, stateAfter) {
     if (stateBefore.lastGlobalSnapshot.startCycle.eq(new BN(this.cycle))) {
         stateAfter.lastGlobalSnapshotIndex.should.be.bignumber.equal(stateBefore.lastGlobalSnapshotIndex); // no new snapshot
     } else {
@@ -33,12 +39,12 @@ const shouldUpdateHistory = async function (receipt, eventName, staker, tokenIds
     let newStakerStake = stateBefore.lastStakerSnapshot.stake;
 
     for (var index = 0; index < tokenIds.length; index++) {
-        if (eventName == 'NftStaked') {
+        if (action == 'stake') {
             newGlobalStake = newGlobalStake.add(stateAfter.tokenInfos[index].weight);
             newStakerStake = newStakerStake.add(stateAfter.tokenInfos[index].weight);
             stateBefore.tokenInfos[index].owner.should.equal(constants.ZeroAddress);
             stateAfter.tokenInfos[index].owner.should.equal(staker);
-        } else if ('NftUnstaked') {
+        } else if (action == 'unstake') {
             newGlobalStake = newGlobalStake.sub(stateAfter.tokenInfos[index].weight);
             newStakerStake = newStakerStake.sub(stateAfter.tokenInfos[index].weight);
             stateBefore.tokenInfos[index].owner.should.equal(staker);
@@ -48,17 +54,6 @@ const shouldUpdateHistory = async function (receipt, eventName, staker, tokenIds
 
     stateAfter.lastGlobalSnapshot.stake.should.be.bignumber.equal(newGlobalStake);
     stateAfter.lastStakerSnapshot.stake.should.be.bignumber.equal(newStakerStake);
-
-    for (var index = 0; index < tokenIds.length; index++) {
-        var tokenId = tokenIds[index];
-
-        await expectEvent.inTransaction(receipt.tx, this.stakingContract, eventName, {
-            staker,
-            cycle: new BN(this.cycle),
-            tokenId,
-            weight: stateAfter.tokenInfos[index].weight,
-        });
-    }
 
     await expectEvent.inTransaction(receipt.tx, this.stakingContract, 'HistoriesUpdated', {
         staker,
@@ -97,8 +92,43 @@ const retrieveStakingState = async function (staker, tokenIds) {
     return state;
 };
 
-const shouldRevertAndNotStakeNft = function (staker, tokenId, error) {
-    it(`[stakeNft] revert and not stake ${tokenId} by ${staker}`, async function () {
+const getTokenIds = async function (staker, tokens, options) {
+    options = options || {}; // falsey initializes as empty options
+
+    const numTokens = tokens.length;
+    const tokenIds = [];
+
+    for (let index = 0; index < numTokens; index++) {
+        tokenIds.push(await getTokenId.bind(this)(staker, tokens[index], options[index]));
+    }
+
+    return tokenIds;
+};
+
+const getTokenId = async function (staker, token, options) {
+    options = options || {}; // falsey initializes as empty options
+
+    const owner = options.owner || staker;
+
+    if (typeof token == 'number' || token instanceof Number) {
+        return this.stakerTokens[owner][token];
+    }
+
+    if (typeof token == 'string' || token instanceof String) {
+        if (options.minter) {
+            await this.nftContract.mintNonFungible(owner, token, {from: minter});
+        }
+
+        return token;
+    }
+};
+
+const shouldRevertAndNotStakeNft = function (staker, token, error, options = {}) {
+    it('shouldRevertAndNotStakeNft', async function () {
+        const tokenId = await getTokenId.bind(this)(staker, token, options);
+
+        this.test.title = `[stakeNft] revert and not stake ${tokenId} by ${staker}`;
+
         await expectRevert(
             this.nftContract.transferFrom(staker, this.stakingContract.address, tokenId, {from: staker}),
             error
@@ -106,17 +136,24 @@ const shouldRevertAndNotStakeNft = function (staker, tokenId, error) {
     });
 };
 
-const shouldRevertAndNotUnstakeNft = function (staker, tokenId, error) {
-    it(`[unstakeNft] revert and not unstake ${tokenId} by ${staker}`, async function () {
+const shouldRevertAndNotUnstakeNft = function (staker, token, error, options = {}) {
+    it('shouldRevertAndNotUnstakeNft', async function () {
+        const tokenId = await getTokenId.bind(this)(staker, token, options);
+
+        this.test.title = `[unstakeNft] revert and not unstake ${tokenId} by ${staker}`;
+
         await expectRevert(this.stakingContract.unstakeNft(tokenId, {from: staker}), error);
     });
 };
 
-const shouldRevertAndNotBatchStakeNfts = function (staker, tokenIds, error) {
-    it(`[stakeNft] revert and not stake ${JSON.stringify(tokenIds)} by ${staker}`, async function () {
+const shouldRevertAndNotBatchStakeNfts = function (staker, tokens, error, options = {}) {
+    it('shouldRevertAndNotBatchStakeNfts', async function () {
+        const tokenIds = await getTokenIds.bind(this)(staker, tokens, options);
+
+        this.test.title = `[stakeNft] revert and not batch stake ${JSON.stringify(tokenIds)} by ${staker}`;
+
         var data = constants.EmptyByte;
         var values = Array(tokenIds.length).fill(1);
-
         await expectRevert(
             this.nftContract.safeBatchTransferFrom(staker, this.stakingContract.address, tokenIds, values, data, {
                 from: staker,
@@ -126,8 +163,20 @@ const shouldRevertAndNotBatchStakeNfts = function (staker, tokenIds, error) {
     });
 };
 
-const shouldStakeNft = function (staker, tokenId) {
-    it(`[stakeNft] ${tokenId} by ${staker}`, async function () {
+const shouldRevertAndNotBatchUnstakeNfts = function (staker, tokens, error, options = {}) {
+    it('shouldRevertAndNotBatchUnstakeNfts', async function () {
+        const tokenIds = await getTokenIds.bind(this)(staker, tokens, options);
+        this.test.title = `[unstakeNft] revert and not batch unstake ${JSON.stringify(tokenIds)} by ${staker}`;
+        await expectRevert(this.stakingContract.batchUnstakeNfts(tokenIds, {from: staker}), error);
+    });
+};
+
+const shouldStakeNft = function (staker, token, options = {}) {
+    it('shouldStakeNft', async function () {
+        const tokenId = await getTokenId.bind(this)(staker, token, options);
+
+        this.test.title = `[stakeNft] ${tokenId} by ${staker}`;
+
         const stateBefore = await retrieveStakingState.bind(this)(staker, [tokenId]);
         const receipt = await this.nftContract.transferFrom(staker, this.stakingContract.address, tokenId, {
             from: staker,
@@ -135,24 +184,45 @@ const shouldStakeNft = function (staker, tokenId) {
         if (this.debug) await debugCurrentState.bind(this)();
         const stateAfter = await retrieveStakingState.bind(this)(staker, [tokenId]);
 
-        await shouldUpdateHistory.bind(this)(receipt, 'NftStaked', staker, [tokenId], stateBefore, stateAfter);
+        await shouldUpdateHistory.bind(this)(receipt, 'stake', staker, [tokenId], stateBefore, stateAfter);
         await shouldSetNextClaimIfUnset.bind(this)(stateBefore, stateAfter);
+
+        await expectEvent.inTransaction(receipt.tx, this.stakingContract, 'NftStaked', {
+            staker,
+            cycle: new BN(this.cycle),
+            tokenId,
+            weight: stateAfter.tokenInfos[0].weight,
+        });
     });
 };
 
-const shouldUnstakeNft = function (staker, tokenId) {
-    it(`[unstakeNft] ${tokenId} by ${staker}`, async function () {
+const shouldUnstakeNft = function (staker, token, options = {}) {
+    it('shouldUnstakeNft', async function () {
+        const tokenId = await getTokenId.bind(this)(staker, token, options);
+        this.test.title = `[unstakeNft] ${tokenId} by ${staker}`;
+
         const stateBefore = await retrieveStakingState.bind(this)(staker, [tokenId]);
         const receipt = await this.stakingContract.unstakeNft(tokenId, {from: staker});
         if (this.debug) await debugCurrentState.bind(this)();
         const stateAfter = await retrieveStakingState.bind(this)(staker, [tokenId]);
 
-        await shouldUpdateHistory.bind(this)(receipt, 'NftUnstaked', staker, [tokenId], stateBefore, stateAfter);
+        await shouldUpdateHistory.bind(this)(receipt, 'unstake', staker, [tokenId], stateBefore, stateAfter);
+
+        await expectEvent.inTransaction(receipt.tx, this.stakingContract, 'NftUnstaked', {
+            staker,
+            cycle: new BN(this.cycle),
+            tokenId,
+            weight: stateAfter.tokenInfos[0].weight,
+        });
     });
 };
 
-const shouldBatchStakeNfts = function (staker, tokenIds) {
-    it(`[stakeNft] ${JSON.stringify(tokenIds)} by ${staker}`, async function () {
+const shouldBatchStakeNfts = function (staker, tokens, options = {}) {
+    it('shouldBatchStakeNfts', async function () {
+        const tokenIds = await getTokenIds.bind(this)(staker, tokens, options);
+
+        this.test.title = `[stakeNft] ${JSON.stringify(tokenIds)} by ${staker}`;
+
         var data = constants.EmptyByte;
         var values = Array(tokenIds.length).fill(1);
 
@@ -168,16 +238,69 @@ const shouldBatchStakeNfts = function (staker, tokenIds) {
         if (this.debug) await debugCurrentState.bind(this)();
         const stateAfter = await retrieveStakingState.bind(this)(staker, tokenIds);
 
-        await shouldUpdateHistory.bind(this)(receipt, 'NftStaked', staker, tokenIds, stateBefore, stateAfter);
+        await shouldUpdateHistory.bind(this)(receipt, 'stake', staker, tokenIds, stateBefore, stateAfter);
         await shouldSetNextClaimIfUnset.bind(this)(stateBefore, stateAfter);
+
+        await expectEvent.inTransaction(receipt.tx, this.stakingContract, 'NftsBatchStaked', {
+            staker,
+            cycle: new BN(this.cycle),
+            tokenIds,
+            weights: stateAfter.tokenInfos.map((t) => t.weight.toString()),
+        });
     });
+};
+
+const shouldBatchUnstakeNfts = function (staker, tokens, options = {}) {
+    it('shouldBatchUnstakeNfts', async function () {
+        const tokenIds = await getTokenIds.bind(this)(staker, tokens, options);
+
+        this.test.title = `[unstakeNft] ${JSON.stringify(tokenIds)} by ${staker}`;
+
+        const stateBefore = await retrieveStakingState.bind(this)(staker, tokenIds);
+        const receipt = await this.stakingContract.batchUnstakeNfts(tokenIds, {from: staker});
+        if (this.debug) await debugCurrentState.bind(this)();
+        const stateAfter = await retrieveStakingState.bind(this)(staker, tokenIds);
+
+        await shouldUpdateHistory.bind(this)(receipt, 'unstake', staker, tokenIds, stateBefore, stateAfter);
+
+        await expectEvent.inTransaction(receipt.tx, this.stakingContract, 'NftsBatchUnstaked', {
+            staker,
+            cycle: new BN(this.cycle),
+            tokenIds,
+            weights: stateAfter.tokenInfos.map((t) => t.weight.toString()),
+        });
+    });
+};
+
+const mintStakerTokens = async function (...stakers) {
+    this.stakerTokens = this.stakerTokens || {};
+
+    const stakerTokens = {};
+    const rarities = RarityWeights.map((item) => item.rarity);
+
+    for (const staker of stakers) {
+        const tokens = [];
+
+        for (const rarity of rarities) {
+            const tokenId = TokenHelper.makeTokenId(rarity, TokenHelper.Types.Car);
+            tokens.push(tokenId);
+            await this.nftContract.mintNonFungible(staker, tokenId, {from: creator});
+        }
+
+        stakerTokens[staker] = tokens;
+    }
+
+    this.stakerTokens = Object.assign(this.stakerTokens, stakerTokens);
 };
 
 module.exports = {
     shouldStakeNft,
     shouldUnstakeNft,
     shouldBatchStakeNfts,
+    shouldBatchUnstakeNfts,
     shouldRevertAndNotStakeNft,
     shouldRevertAndNotUnstakeNft,
     shouldRevertAndNotBatchStakeNfts,
+    shouldRevertAndNotBatchUnstakeNfts,
+    mintStakerTokens,
 };